Resolving issue with Fixes AML
Previously the fixes AML was trying to allow users to select the DQD and PQD ItemTypes as deliverables of the new project template being added. However, the List that the fixes AML was modifying is actually just a dynamically created List based on the poly sources of the Deliverable Poly ItemType. With this new information, I'm updating the fixes AML to instead edit the Deliverable ItemType to add new poly sources for the DQD and PQD ItemTypes. This should add the values to the list as well as handle any other functionality that relies on these deliverables being actual poly sources of the Deliverable poly ItemType.